Move Organisation Feed classes into module
Co-locating these files makes it easier to understand what's related.
