+++ app/services/feeds/organisation_feed_content.rb
@@ -1,0 +1,34 @@
+module Feeds
+  class OrganisationFeedContent
+    include RummagerFields
+
+    attr_reader :organisation
+
+    def initialize(organisation:)
+      @organisation = organisation
+    end
+
+    def self.fetch(organisation:)
+      new(organisation: organisation).fetch
+    end
+
+    def fetch
+      search_response
+    end
+
+  private
+
+    def search_response
+      params = {
+        start: 0,
+        count: 20,
+        fields: RummagerFields::FEED_SEARCH_FIELDS,
+        filter_organisations: organisation,
+        reject_content_purpose_supergroup: "other",
+        order: '-public_timestamp',
+      }
+
+      Services.rummager.search(params)
+    end
+  end
+end
